demo-langgraph-workflow/requirements.txt
nats-py==2.7.2
redis==5.0.1
asyncio-nats-client==0.11.5
